// src/shared/api/openapi.d.ts
export interface components {
    schemas: {
        SignInRequest: {
            /** Format: email */
            email: string;
            password: string;
        };
        AuthTokenResponse: {
            /** @description JWT token. Decoded payload contains id and exp. */
            accessToken: string;
            /** @description JWT token. Decoded payload contains id and exp. */
            refreshToken: string;
        };
        UserResponse: {
            name: string;
            memo: string;
        };
        DashboardResponse: {
            numOfTask: number;
            numOfRestTask: number;
            numOfDoneTask: number;
        };
        TaskItem: {
            id: string;
            title: string;
            memo: string;
            /** @enum {string} */
            status: "TODO" | "DONE";
        };
        TaskListResponse: {
            data: components["schemas"]["TaskItem"][];
            hasNext: boolean;
        };
        TaskDetailResponse: {
            title: string;
            memo: string;
            /** Format: date-time */
            registerDatetime: string;
        };
        DeleteTaskResponse: {
            /** @constant */
            success: true;
        };
        ErrorResponse: {
            errorMessage: string;
        };
    };
    responses: never;
    parameters: {
        TaskIdPath: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}